refactor: devided list.h into container/buff.h and container/list.h
- list.h was removed and was devided into the following files:
  - container/buff.h
  - container/buff/class.h
  - container/buff/prototype.h
  - container/buff/generics.h
  - container/list.h
  - container/list/class.h
  - container/list/prototype.h
  - container/list/generics.h

- list.hpp was removed and was integrated the above files.
