+++ include/cparsec2/container/base.h
@@ -1,0 +1,9 @@
+/* -*- coding:utf-8-unix -*- */
+#pragma once
+
+#include <cparsec2/alloc.h>
+#include <cparsec2/codegen.h>
+#include <cparsec2/macros.h>
+#include <cparsec2/none.h>
+
+#define ELEMENT_TYPE(T) CAT(T, _elem_type)
