+++ include/cparsec2/container/buff.h
@@ -1,0 +1,11 @@
+/* -*- coding:utf-8-unix -*- */
+#pragma once
+
+#include "base.h"
+
+#include "list.h"
+
+#include "buff/class.h"
+#include "buff/prototype.h"
+
+#include "buff/generics.h"
