+++ include/cparsec2/container/list.h
@@ -1,0 +1,9 @@
+/* -*- coding:utf-8-unix -*- */
+#pragma once
+
+#include "base.h"
+
+#include "list/class.h"
+#include "list/prototype.h"
+
+#include "list/generics.h"
