+++ src/stream.c
@@ -8,7 +8,7 @@
 #include <string.h>
 
 #include <cparsec2/alloc.h>
-#include <cparsec2/list.h>
+#include <cparsec2/container.h>
 
 struct stStream {
   const char* beg;
